+++ Source/j1WinScene.cpp
@@ -72,7 +72,7 @@ bool j1WinScene::Update(float dt)
 	int h = App->win->GetHeight();
 
 	App->render->Blit(background, 0, 0, NULL, true, App->render->renderer, 1.5);
-	App->render->Blit(wintext, w * 1.55f, h, NULL, true, App->render->renderer, 0.2);
+	App->render->Blit(wintext, (int)(w * 1.55f), h, NULL, true, App->render->renderer, 0.2f);
 
 	//Need to create a timer
 	/*if (App->input->GetKey(SDL_SCANCODE_RETURN)) {
